backend/requirements.txt
uvicorn
fastapi
llama-index==0.12.0
websockets
# llama-index-llms-huggingface==0.4.0
# llama-index-embeddings-huggingface==0.4.0

llama-index-embeddings-huggingface==0.4.0
llama-index-llms-llama-cpp==0.3.0
python-dotenv==1.0.1
llama-index-llms-ollama==0.4.0
llama-cpp-python
langchain==0.3.7
langchain_community==0.3.7
